# docs/api/swagger.yaml
basePath: /
definitions:
...
  presenters.UserCreate:
    properties:
      name:
        type: string
      password:
        type: string
      username:
        type: string
    type: object
  presenters.UserCredentials:
    properties:
      password:
        type: string
      username:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: 975316@telkom.co.id
    name: API Support
  description: BigVision Operation
  license:
    name: MIT
    url: https://opensource.org/license/mit
  termsOfService: http://swagger.io/terms/
  title: BigVision Operation API
  version: "1.0"
paths:
...
  /api/user/login:
    post:
      consumes:
      - application/json
      description: Authenticate user and retrieve JWT token
      parameters:
      - description: User login payload
        in: body
        name: UserCredentials
        required: true
        schema:
          $ref: '#/definitions/presenters.UserCredentials'
      produces:
      - application/json
      responses: {}
      summary: Login user
      tags:
      - User
  /api/user/register:
    post:
      consumes:
      - application/json
      description: Create a new user account
      parameters:
      - description: User registration payload
        in: body
        name: UserCreate
        required: true
        schema:
          $ref: '#/definitions/presenters.UserCreate'
      produces:
      - application/json
      responses: {}
      summary: Register a new user
      tags:
      - User
...
security:
- BearerAuth: []
securityDefinitions:
  BearerAuth:
    description: '"JWT Bearer token required for authentication"'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"